% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_to_df.R
\name{article_to_df}
\alias{article_to_df}
\title{Cast PubMed Data into a List of Articles}
\usage{
article_to_df(
  pubmedArticle,
  autofill = FALSE,
  max_chars = 500,
  getKeywords = FALSE,
  getAuthors = TRUE
)
}
\arguments{
\item{pubmedArticle}{String including one PubMed record.}

\item{autofill}{Logical. If TRUE, missing affiliations are automatically imputed based on other non-NA 
addresses from the same record.}

\item{max_chars}{Numeric (integer). Maximum number of characters to be extracted from the Article 
Abstract field. Set max_chars to -1 for extracting the full-length abstract. Set max_chars to 0 to 
extract no abstract.}

\item{getKeywords}{Logical. If TRUE, an attempt to extract article Keywords will be made.}

\item{getAuthors}{Logical. If FALSE, author information won't be extracted. This will considerably 
speed up the operation.}
}
\value{
Data frame including the extracted features. Each row correspond a different author.
}
\description{
Convert an XML object of PubMed records into a list of strings 
(character vector of length 1) corresponding to individual PubMed articles. 
PubMed records are identified by a "/PubmedArticle" XML tag. This automatically casts 
all the content of each PubMed record to a character-class object without removing XML tags.
}
\details{
Given one Pubmed Article record, this function will automatically extract a set of features. 
Extracted information include: PMID, DOI, article title, article abstract, publication date (year, month, day), 
journal name (title, abbreviation), keywords, and a set of author-specific info (names, affiliation, email address). 
Each row of the output data frame corresponds to one of the authors of the PubMed record. Author-independent info 
(publication ID, title, journal, date) are identical across all rows. If information about authors are not required, 
set 'getAuthors' = TRUE.
}
\examples{
\dontrun{
## Retrieve PubMed data and return a list ot articles
Articles_List <- articles_to_list("pubmed_result.xml")
Articles_DF <- map_dfr(Articles_List, article_to_df) 
}

}
\references{
\url{https://www.data-pulse.com/dev_site/easypubmed/}
}
\author{
Damiano Fantini \email{damiano.fantini@gmail.com}
}
