% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles_to_list.R
\name{articles_to_list}
\alias{articles_to_list}
\title{Cast PubMed Data into a List of Articles}
\usage{
articles_to_list(pubmed_data)
}
\arguments{
\item{pubmed_data}{String corresponding to the name of an XML file 
(typically, the result of a batch_pubmed_download() call). Alternatively, 
an XML Object, such as the result of a fetch_pubmed_data() call.}
}
\description{
Convert an XML object of PubMed records into a list of strings 
(character vector of length 1) corresponding to individual PubMed articles. 
PubMed records are identified by a "/PubmedArticle" XML tag. This automatically casts 
all the content of each PubMed record to a character-class object without removing XML tags.
}
\details{
The input is an XML object or an XML file, typically the result of a fetch_pubmed_data() 
call or a batch_pubmed_download() call. The function returns a list where each element 
is a different PubMed record

Character vector listing all the records from the original XML object in text format. Elements in the list are not named and are only accessible via the numeric index.
}
\examples{
\dontrun{
## Retrieve PubMed data and return a list ot articles
listed_articles <- articles_to_list("pubmed_result.xml")
}

}
\references{
https://www.data-pulse.com/dev_site/easypubmed/
}
\author{
Damiano Fantini <damiano.fantini@gmail.com>
}
