% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_patent_num.R
\name{extract_patent_num}
\alias{extract_patent_num}
\title{Method to get patents data based on a web search}
\usage{
extract_patent_num(url_search, nb_pages)
}
\arguments{
\item{url_search}{Url of the search}

\item{nb_pages}{Number of pages to parse (number of patents found / 50)}
}
\value{
A list with results
}
\description{
Method to get patents data based on a web search
}
\examples{
\dontrun{
get_pubmed_df_general('query')
}

}
