% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_applications_df.R
\name{get_applications_df}
\alias{get_applications_df}
\title{Extract applications information as a tibble}
\usage{
get_applications_df(pv_res)
}
\arguments{
\item{pv_res}{Takes results from get_patents_data}
}
\value{
A tibble with the applications information
}
\description{
Extract applications information as a tibble
}
\examples{
\dontrun{
get_applications_df(pv_res)
}

}
