% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignees_df.R
\name{get_assignees_df}
\alias{get_assignees_df}
\title{Extract assignees information as a tibble}
\usage{
get_assignees_df(pv_res)
}
\arguments{
\item{pv_res}{Takes results from get_patents_data}
}
\value{
A tibble with the assignees information
}
\description{
Extract assignees information as a tibble
}
\examples{
\dontrun{
get_assignees_df(pv_res)
}

}
