% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpcs_df.R
\name{get_cpcs_df}
\alias{get_cpcs_df}
\title{Extract cpcs information as a tibble}
\usage{
get_cpcs_df(pv_res)
}
\arguments{
\item{pv_res}{Takes results from get_patents_data}
}
\value{
A tibble with the cpcs information
}
\description{
Extract cpcs information as a tibble
}
\examples{
\dontrun{
get_cpcs_df(pv_res)
}

}
