% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_examiners_df.R
\name{get_examiners_df}
\alias{get_examiners_df}
\title{Extract examiners information as a tibble}
\usage{
get_examiners_df(pv_res)
}
\arguments{
\item{pv_res}{Takes results from get_patents_data}
}
\value{
A tibble with the examiners information
}
\description{
Extract examiners information as a tibble
}
\examples{
\dontrun{
get_examiners_df(pv_res)
}

}
