% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inventors_df.R
\name{get_inventors_df}
\alias{get_inventors_df}
\title{Extract inventors information as a tibble}
\usage{
get_inventors_df(pv_res)
}
\arguments{
\item{pv_res}{Takes results from get_patents_data}
}
\value{
A tibble with the inventors information
}
\description{
Extract inventors information as a tibble
}
\examples{
\dontrun{
get_inventors_df(pv_res)
}

}
