% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lawyers_df.R
\name{get_lawyers_df}
\alias{get_lawyers_df}
\title{Extract lawyers information as a tibble}
\usage{
get_lawyers_df(pv_res)
}
\arguments{
\item{pv_res}{Takes results from get_patents_data}
}
\value{
A tibble with the lawyers information
}
\description{
Extract lawyers information as a tibble
}
\examples{
\dontrun{
get_lawyers_df(pv_res)
}

}
