% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patents_data.R
\name{get_patents_data}
\alias{get_patents_data}
\title{Fetch USPTO patents data}
\usage{
get_patents_data(query)
}
\arguments{
\item{query}{Takes in a query formated with `patentsview` domain specific language (DSL)}
}
\value{
List containing patents information
}
\description{
Fetch USPTO patents data
}
\examples{
\dontrun{
query <- with_qfuns( 
 or(
  text_all(patent_title = "artificial intelligence"),
  text_all(patent_abstract = "artificial intelligence")
  )
)
get_patents_data(query)
}

}
