% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patents_df.R
\name{get_patents_df}
\alias{get_patents_df}
\title{Extract patents information as a tibble}
\usage{
get_patents_df(pv_res)
}
\arguments{
\item{pv_res}{Takes results from get_patents_data}
}
\value{
A tibble with the patents information
}
\description{
Extract patents information as a tibble
}
\examples{
\dontrun{
get_patents_df(pv_res)
}

}
