% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pubmed_df.R
\name{get_pubmed_df}
\alias{get_pubmed_df}
\title{Perform search on pudmed with RISmed and return a tibble with infos}
\usage{
get_pubmed_df(query)
}
\arguments{
\item{query}{A pubmed query}
}
\value{
A tibble with scientific publication informations
}
\description{
Perform search on pudmed with RISmed and return a tibble with infos
}
\examples{
\dontrun{
get_examiners_df(pv_res)
}

}
