% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pubmed_df.R
\name{get_pubmed_df_abstract}
\alias{get_pubmed_df_abstract}
\title{Perform search on pudmed with RISmed and return a tibble with abstracts}
\usage{
get_pubmed_df_abstract(query)
}
\arguments{
\item{query}{A pubmed query}
}
\value{
A tibble with scientific publication informations
}
\description{
Perform search on pudmed with RISmed and return a tibble with abstracts
}
\examples{
\dontrun{
get_pubmed_df_abstract('query')
}

}
