% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wipos_df.R
\name{get_wipos_df}
\alias{get_wipos_df}
\title{Extract wipos information as a tibble}
\usage{
get_wipos_df(pv_res)
}
\arguments{
\item{pv_res}{Takes results from get_patents_data}
}
\value{
A tibble with the wipos information
}
\description{
Extract wipos information as a tibble
}
\examples{
\dontrun{
get_wipos_df(pv_res)
}

}
