% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pubmed_XML.R
\name{read_pubmed_XML}
\alias{read_pubmed_XML}
\title{Parse pubmed XML file with easyPubMed package}
\usage{
read_pubmed_XML(xml_file)
}
\arguments{
\item{xml_file}{An XML file obtained from pubmed}
}
\value{
A data frame
}
\description{
Parse pubmed XML file with easyPubMed package
}
